fix: fix @cf_skip
